[#816] Moved utils.js into utilities namespace
