+++ client/src/components/modals/InviteUser.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import Modal from 'react-modal';
-import isValidEmail from '../../utils';
+import isValidEmail from '../../utilities/validators';
 
 require('../DashboardModal.scss');
